libgimp: change signature of get_window() of GimpProgressVtable.
As diagnosed in #8649, using a guint32 for windows identifier may have been
right long ago (was it?), but is definitely not anymore. I can see that a XID is
an unsigned long nowadays (usually 64-bit on 64-bit Linux).

As far as I can see, on Windows, it would be a void* behind (which also
corresponds to the error message in #8649 description):

> typedef void *PVOID;
> typedef PVOID HANDLE;
> typedef HANDLE HWND;

Cf. https://learn.microsoft.com/en-us/windows/win32/winprog/windows-data-types

I *think* that pointers would be 64-bit on Windows 64-bit, though I'm unsure
(after all, this is an OS with 32-bit long int on 64-bit arch!).

Anyway, it's just better to move to 64-bit window identifiers.

+++ libgimp/gimpprogress.c
@@ -382,7 +382,7 @@ gimp_temp_progress_run (GimpProcedure        *procedure,
     case GIMP_PROGRESS_COMMAND_GET_WINDOW:
       {
         GimpValueArray *return_vals;
-        guint32         window_id = 0;
+        guint64         window_id = 0;
 
         if (progress_data->vtable.get_window)
           window_id = progress_data->vtable.get_window (progress_data->data);
